Fix some bugs

+++ src/more_knowledge/unit_int.cpp
@@ -4,34 +4,34 @@
 using namespace std;
 
 int method1(int n) {
-  int i = 10;
-  int c = 0;
+	int i = 10;
+	int c = 0;
 
-  while ((c = n % i) != n) {
-    n -= c;
-    i *= 10;
-  }
-  
-  return n;
+	while ((c = n % i) != n) {
+		n -= c;
+		i *= 10;
+	}
+
+	return n;
 }
 
 // Get counts of digits
 // Complexity: constant
 int getCountsOfDigits(int n) {
-  return n == 0 ? 1 : ceil(log10(abs(n) + 0.5));
+	return n == 0 ? 1 : (int)ceil(log10(abs(n) + 0.5));
 }
 
 int method2(int n) {
-  int c = getCountsOfDigits(n);
-  int d = pow(10, c - 1);
-  return n - (n % d);
+	int c = getCountsOfDigits(n);
+	int d = (int)pow(10, c - 1);
+	return n - (n % d);
 }
 
 int main(int argc, char *argv[]) {
-  const int n = 835047;
+	const int n = 835047;
 
-  cout << "Method 1: " << method1(n) << endl; // 800000
-  cout << "Method 2: " << method2(n) << endl; // 800000
+	cout << "Method 1: " << method1(n) << endl; // 800000
+	cout << "Method 2: " << method2(n) << endl; // 800000
 
-  return 0;
+	return 0;
 }
